pyproject.toml
[project]
name = "umet-reports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.4",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "mysql-connector-python>=9.3.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "tenacity>=9.1.2",
]
